backend/requirements.txt
flask >= 1.1.1
PyJWT >= 1.7.1
pymongo[srv] >= 3.9.0
